requirements.txt
numpy>=1.18.0
pandas>=1.0.0
scikit-learn>=0.22.0
matplotlib>=3.1.0
seaborn>=0.10.0
joblib>=0.14.0
Flask>=2.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
dill>=0.3.4
mlflow>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
optuna>=3.0.0
apache-airflow>=2.7.0